Recalculate fixes
